refactor(cli) applytemplate
